+++ SharpSource/SharpSource.Test/StaticInitializerAccessedBeforeInitializationTests.cs
@@ -341,7 +341,7 @@ struct Test
         await VerifyDiagnostic(original, "FirstField accesses SecondField but both are marked as static and SecondField will not be initialized when it is used");
     }
 
-    [BugVerificationTest(IssueUrl = "https://github.com/Vannevelj/SharpSource/issues/172")]
+    [BugVerificationTest(IssueUrl = "https://github.com/Vannevelj/SharpSource/issues/171")]
     public async Task StaticInitializerAccessedBeforeInitialization_MethodInvocation()
     {
         var original = @"
@@ -366,5 +366,35 @@ class Test
 }";
 
         await VerifyDiagnostic(original, "FirstField accesses SomeArg but both are marked as static and SomeArg will not be initialized when it is used");
+    }
+
+    [BugVerificationTest(IssueUrl = "https://github.com/Vannevelj/SharpSource/issues/173")]
+    public async Task StaticInitializerAccessedBeforeInitialization_Action()
+    {
+        var original = @"
+using System;
+
+class Test
+{
+	static Action<int> FirstField = DoThing;
+	static void DoThing(int i) { }
+}";
+
+        await VerifyDiagnostic(original);
+    }
+
+    [BugVerificationTest(IssueUrl = "https://github.com/Vannevelj/SharpSource/issues/173")]
+    public async Task StaticInitializerAccessedBeforeInitialization_Func()
+    {
+        var original = @"
+using System;
+
+class Test
+{
+	static Func<int> FirstField = DoThing;
+	static int DoThing() => 32;
+}";
+
+        await VerifyDiagnostic(original);
     }
 }