+++ CHANGELOG.md
@@ -4,6 +4,7 @@ https://keepachangelog.com/en/1.0.0/
 ## [1.9.2] - 2022-09-12
 - `StaticInitializerAccessedBeforeInitialization`: now takes `nameof()` usage into account
 - `StaticInitializerAccessedBeforeInitialization`: no longer triggers for invocations of `static` functions
+- `StaticInitializerAccessedBeforeInitialization`: no longer triggers when the field is of type `Action` or `Func`
 
 ## [1.9.1] - 2022-09-12
 - Internal code cleanup to remove -Async suffixes on tests